[runtime] Use pthread_cond_timedwait_relative_np () function on osx if available.
