+++ mono/utils/Makefile.am
@@ -59,6 +59,7 @@ monoutils_sources = \
 	mono-mmap-internals.h	\
 	mono-mmap-windows-internals.h	\
 	mono-os-mutex.h		\
+	mono-os-mutex.c		\
 	mono-coop-mutex.h		\
 	mono-once.h		\
 	mono-lazy-init.h		\
